+++ packages/iov-rise/src/riseconnection.spec.ts
@@ -1,9 +1,11 @@
+import Long from "long";
 import { ReadonlyDate } from "readonly-date";
 
 import { Algorithm, ChainId, PublicKeyBundle, PublicKeyBytes, SignatureBytes } from "@iov/base-types";
 import {
   Address,
   Amount,
+  BcpAccount,
   BcpAccountQuery,
   BcpAddressQuery,
   BcpBlockInfo,
@@ -15,15 +17,22 @@ import {
   TokenTicker,
   TransactionId,
 } from "@iov/bcp-types";
+import { Random } from "@iov/crypto";
 import { Derivation } from "@iov/dpos";
 import { Encoding } from "@iov/encoding";
 import { Ed25519Wallet } from "@iov/keycontrol";
 
+import { pubkeyToAddress } from "./derivation";
 import { riseCodec } from "./risecodec";
 import { generateNonce, RiseConnection } from "./riseconnection";
 
 const { fromHex } = Encoding;
 const riseTestnet = "e90d39ac200c495b97deb6d9700745177c7fc4aa80a404108ec820cbeced054c" as ChainId;
+
+async function randomAddress(): Promise<Address> {
+  const pubkey = await Random.getBytes(32);
+  return pubkeyToAddress(pubkey);
+}
 
 describe("RiseConnection", () => {
   const base = "https://twallet.rise.vision";
@@ -207,6 +216,94 @@ describe("RiseConnection", () => {
     }
 
     connection.disconnect();
+  });
+
+  describe("watchAccount", () => {
+    it("can watch account by address", done => {
+      (async () => {
+        const connection = await RiseConnection.establish(base);
+
+        const recipient = await randomAddress();
+
+        const events = new Array<BcpAccount | undefined>();
+        const subscription = connection.watchAccount({ address: recipient }).subscribe({
+          next: event => {
+            events.push(event);
+
+            if (events.length === 3) {
+              const [event1, event2, event3] = events;
+
+              expect(event1).toBeUndefined();
+
+              if (!event2) {
+                throw new Error("Second event must not be undefined");
+              }
+              expect(event2.address).toEqual(recipient);
+              expect(event2.name).toBeUndefined();
+              expect(event2.pubkey).toBeUndefined();
+              expect(event2.balance.length).toEqual(1);
+              expect(event2.balance[0].quantity).toEqual(defaultSendAmount.quantity);
+              expect(event2.balance[0].tokenTicker).toEqual(defaultSendAmount.tokenTicker);
+
+              if (!event3) {
+                throw new Error("Second event must not be undefined");
+              }
+              expect(event3.address).toEqual(recipient);
+              expect(event3.name).toBeUndefined();
+              expect(event3.pubkey).toBeUndefined();
+              expect(event3.balance.length).toEqual(1);
+              expect(event3.balance[0].quantity).toEqual(
+                Long.fromString(defaultSendAmount.quantity)
+                  .multiply(2)
+                  .toString(),
+              );
+              expect(event3.balance[0].tokenTicker).toEqual(defaultSendAmount.tokenTicker);
+
+              subscription.unsubscribe();
+              connection.disconnect();
+              done();
+            }
+          },
+          complete: done.fail,
+          error: done.fail,
+        });
+
+        const wallet = new Ed25519Wallet();
+        const mainIdentity = await wallet.createIdentity(await defaultKeypair);
+
+        for (const _ of [0, 1]) {
+          const sendTx: SendTransaction = {
+            kind: "bcp/send",
+            chainId: riseTestnet,
+            signer: mainIdentity.pubkey,
+            recipient: recipient,
+            amount: defaultSendAmount,
+          };
+
+          const nonce = generateNonce();
+          const signingJob = riseCodec.bytesToSign(sendTx, nonce);
+          const signature = await wallet.createTransactionSignature(
+            mainIdentity,
+            signingJob.bytes,
+            signingJob.prehashType,
+            riseTestnet,
+          );
+
+          const signedTransaction: SignedTransaction = {
+            transaction: sendTx,
+            primarySignature: {
+              nonce: nonce,
+              pubkey: mainIdentity.pubkey,
+              signature: signature,
+            },
+            otherSignatures: [],
+          };
+
+          const result = await connection.postTx(riseCodec.bytesToPost(signedTransaction));
+          await result.blockInfo.waitFor(info => info.state === BcpTransactionState.InBlock);
+        }
+      })().catch(done.fail);
+    }, 90_000);
   });
 
   describe("getBlockHeader", () => {