+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.10.4
 
 * @iov/lisk: Implement `LiskConnection.watchAccount`
+* @iov/rise: Implement `RiseConnection.watchAccount`
 
 ## 0.10.3